[common,tools] Remove copying from Optional EmplaceValue method
This method constructs temporary object and copies it to the current instance.
Event if this branch is not used, it still creates temporary object on stack.
That leads to uncontrolled stack usage in case big object is hold by optional.
Remove copying branch and use only new object construct to avoid stack usage.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>