[sm,launcher] Try to stop only services with runtime dir on start
In case of crash or unexpected SM restart, we may have some service instances
started or partially started. We should try to cleanup resources occupied by
them. If there is no runtime dir for specific instance, this is treated as
instance doesn't hold any resources and no need to try to stop it.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>